Improve LPMUITests (#3796)
## Summary
- Moves LPMUITests class to its own file
- DRYs it up a little bit w/ some helper methods
- Audits each test and makes sure it completes the entire flow, except
for the LPMs that redirect to 3rd party websites (e.g. amazon.com).

- Also cleans up `getTextFieldElement` etc helper methods and makes them
force unwrap. Sorry, kind of unrelated to the rest of this PR.

## Changelog
Not user facing
