+++ Example/PaymentSheet Example/PaymentSheetUITest/PaymentSheetUITest.swift
@@ -445,1176 +445,6 @@ class PaymentSheetStandardUITests: PaymentSheetUITestCase {
     }
 }
 
-class PaymentSheetStandardLPMUITests: PaymentSheetUITestCase {
-    func testEPSPaymentMethodHasTextFieldsAndDropdown() throws {
-        var settings = PaymentSheetTestPlaygroundSettings.defaultValues()
-        settings.customerMode = .new
-        settings.applePayEnabled = .off
-        settings.currency = .eur
-        loadPlayground(app, settings)
-
-        app.buttons["Present PaymentSheet"].tap()
-        let payButton = app.buttons["Pay €50.99"]
-
-        guard let eps = scroll(collectionView: app.collectionViews.firstMatch, toFindCellWithId: "EPS") else {
-            XCTFail()
-            return
-        }
-        eps.tap()
-
-        XCTAssertFalse(payButton.isEnabled)
-        let name = app.textFields["Full name"]
-        name.tap()
-        name.typeText("John Doe")
-        name.typeText(XCUIKeyboardKey.return.rawValue)
-
-        let bank = app.textFields["EPS Bank"]
-        bank.tap()
-        app.pickerWheels.firstMatch.adjust(toPickerWheelValue: "BKS Bank AG")
-        app.toolbars.buttons["Done"].tap()
-
-        payButton.tap()
-
-        let webviewCloseButton = app.otherElements["TopBrowserBar"].buttons["Close"]
-        XCTAssertTrue(webviewCloseButton.waitForExistence(timeout: 10.0))
-        webviewCloseButton.tap()
-    }
-
-    func testP24PaymentMethodHasTextFieldsAndDropdown() throws {
-        var settings = PaymentSheetTestPlaygroundSettings.defaultValues()
-        settings.customerMode = .new
-        settings.applePayEnabled = .off
-        settings.currency = .eur
-        loadPlayground(app, settings)
-
-        app.buttons["Present PaymentSheet"].tap()
-
-        let payButton = app.buttons["Pay €50.99"]
-        guard let p24 = scroll(collectionView: app.collectionViews.firstMatch, toFindCellWithId: "Przelewy24") else {
-            XCTFail()
-            return
-        }
-        p24.tap()
-
-        XCTAssertFalse(payButton.isEnabled)
-        let name = app.textFields["Full name"]
-        name.tap()
-        name.typeText("John Doe")
-        name.typeText(XCUIKeyboardKey.return.rawValue)
-
-        XCTAssertFalse(payButton.isEnabled)
-        let email = app.textFields["Email"]
-        email.tap()
-        email.typeText("test@test.com")
-        email.typeText(XCUIKeyboardKey.return.rawValue)
-
-        let bank = app.textFields["Przelewy24 Bank"]
-        bank.tap()
-        app.pickerWheels.firstMatch.adjust(toPickerWheelValue: "BNP Paribas")
-        app.toolbars.buttons["Done"].tap()
-
-        payButton.tap()
-
-        let webviewCloseButton = app.otherElements["TopBrowserBar"].buttons["Close"]
-        XCTAssertTrue(webviewCloseButton.waitForExistence(timeout: 10.0))
-        webviewCloseButton.tap()
-    }
-
-    // Klarna has a text field and country drop down
-    func testKlarnaPaymentMethod() throws {
-        var settings = PaymentSheetTestPlaygroundSettings.defaultValues()
-        settings.customerMode = .new // new customer
-        settings.apmsEnabled = .off
-        loadPlayground(app, settings)
-
-        app.buttons["Present PaymentSheet"].tap()
-        let payButton = app.buttons["Pay $50.99"]
-
-        // Select Klarna
-        guard let klarna = scroll(collectionView: app.collectionViews.firstMatch, toFindCellWithId: "Klarna") else {
-            XCTFail()
-            return
-        }
-        klarna.tap()
-
-        XCTAssertFalse(payButton.isEnabled)
-        let name = app.textFields["Email"]
-        name.tap()
-        name.typeText("foo@bar.com")
-        name.typeText(XCUIKeyboardKey.return.rawValue)
-
-        // Country should be pre-filled
-
-        // Attempt payment
-        payButton.tap()
-
-        // Klarna uses ASWebAuthenticationSession, tap continue to allow the web view to open:
-        let springboard = XCUIApplication(bundleIdentifier: "com.apple.springboard")
-        let sbContinueButton = springboard.buttons["Continue"]
-        XCTAssertTrue(sbContinueButton.waitForExistence(timeout: 10.0))
-        sbContinueButton.tap()
-
-        let webviewCloseButton = app.buttons["Cancel"]
-        XCTAssertTrue(webviewCloseButton.waitForExistence(timeout: 10.0))
-        webviewCloseButton.tap()
-    }
-
-    func testAffirmPaymentMethod() throws {
-        var settings = PaymentSheetTestPlaygroundSettings.defaultValues()
-        settings.customerMode = .new
-        settings.apmsEnabled = .off
-        loadPlayground(app, settings)
-        app.buttons["Present PaymentSheet"].tap()
-        let payButton = app.buttons["Pay $50.99"]
-
-        // Select affirm
-        guard let affirm = scroll(collectionView: app.collectionViews.firstMatch, toFindCellWithId: "Affirm") else {
-            XCTFail()
-            return
-        }
-        affirm.tap()
-
-        XCTAssertTrue(payButton.isEnabled)
-
-        // Attempt payment, should fail
-        payButton.tap()
-    }
-
-    func testAmazonPayPaymentMethod() throws {
-        var settings = PaymentSheetTestPlaygroundSettings.defaultValues()
-        settings.customerMode = .new
-        settings.apmsEnabled = .off
-        loadPlayground(app, settings)
-        app.buttons["Present PaymentSheet"].tap()
-        let payButton = app.buttons["Pay $50.99"]
-
-        // Select Amazon Pay
-        guard let amazonPay = scroll(collectionView: app.collectionViews.firstMatch, toFindCellWithId: "Amazon Pay") else {
-            XCTFail()
-            return
-        }
-        amazonPay.tap()
-
-        XCTAssertTrue(payButton.isEnabled)
-
-        // Attempt payment, should succeed
-        payButton.tap()
-    }
-
-    func testAlmaPaymentMethod() throws {
-        var settings = PaymentSheetTestPlaygroundSettings.defaultValues()
-        settings.currency = .eur
-        settings.merchantCountryCode = .FR
-        settings.customerMode = .new
-        settings.apmsEnabled = .off
-        loadPlayground(app, settings)
-        app.buttons["Present PaymentSheet"].tap()
-        let payButton = app.buttons["Pay €50.99"]
-
-        // Select Alma
-        guard let alma = scroll(collectionView: app.collectionViews.firstMatch, toFindCellWithId: "Alma") else {
-            XCTFail()
-            return
-        }
-        alma.tap()
-
-        XCTAssertTrue(payButton.isEnabled)
-
-        // Attempt payment, should succeed
-        payButton.tap()
-    }
-
-    func testZipPaymentMethod() throws {
-        var settings = PaymentSheetTestPlaygroundSettings.defaultValues()
-        settings.customerMode = .new // new customer
-        settings.apmsEnabled = .off
-        settings.currency = .aud
-        settings.merchantCountryCode = .AU
-        loadPlayground(app, settings)
-
-        app.buttons["Present PaymentSheet"].tap()
-        let payButton = app.buttons["Pay A$50.99"]
-
-        // Select Zip
-        guard let zip = scroll(collectionView: app.collectionViews.firstMatch, toFindCellWithId: "Zip")
-        else {
-            XCTFail()
-            return
-        }
-        zip.tap()
-
-        // Attempt payment
-        payButton.tap()
-
-        // Close the webview, no need to see the successful pay
-        let webviewCloseButton = app.otherElements["TopBrowserBar"].buttons["Close"]
-        XCTAssertTrue(webviewCloseButton.waitForExistence(timeout: 10.0))
-        webviewCloseButton.tap()
-    }
-
-    func testCashAppPaymentMethod() throws {
-        var settings = PaymentSheetTestPlaygroundSettings.defaultValues()
-        settings.customerMode = .new
-        settings.apmsEnabled = .on
-        loadPlayground(app, settings)
-
-        app.buttons["Present PaymentSheet"].tap()
-        let payButton = app.buttons["Pay $50.99"]
-
-        // Select Cash App
-        guard let cashApp = scroll(collectionView: app.collectionViews.firstMatch, toFindCellWithId: "Cash App Pay")
-        else {
-            XCTFail()
-            return
-        }
-        cashApp.tap()
-
-        // Attempt payment
-        payButton.tap()
-
-        // Close the webview, to simulate cancel
-        let webviewCloseButton = app.otherElements["TopBrowserBar"].buttons["Close"]
-        XCTAssertTrue(webviewCloseButton.waitForExistence(timeout: 10.0))
-        webviewCloseButton.tap()
-
-        // Tap to attempt a payment, but fail it
-        payButton.waitForExistenceAndTap()
-        let failPaymentText = app.firstDescendant(withLabel: "FAIL TEST PAYMENT")
-        failPaymentText.waitForExistenceAndTap(timeout: 15.0)
-
-        XCTAssertTrue(app.staticTexts["The customer declined this payment."].waitForExistence(timeout: 5.0))
-
-        // Tap to attempt a payment
-        payButton.waitForExistenceAndTap()
-        let approvePaymentText = app.firstDescendant(withLabel: "AUTHORIZE TEST PAYMENT")
-        approvePaymentText.waitForExistenceAndTap(timeout: 15.0)
-
-        XCTAssertTrue(app.staticTexts["Success!"].waitForExistence(timeout: 15.0))
-    }
-
-    func testCashAppPaymentMethod_setup() throws {
-        var settings = PaymentSheetTestPlaygroundSettings.defaultValues()
-        settings.customerMode = .new
-        settings.apmsEnabled = .on
-        settings.mode = .setup
-        loadPlayground(
-            app,
-            settings
-        )
-
-        app.buttons["Present PaymentSheet"].tap()
-        let setupButton = app.buttons["Set up"]
-
-        // Select Cash App
-        guard let cashApp = scroll(collectionView: app.collectionViews.firstMatch, toFindCellWithId: "Cash App Pay")
-        else {
-            XCTFail()
-            return
-        }
-        cashApp.tap()
-
-        // Attempt set up
-        setupButton.tap()
-
-        // Close the webview, to simulate cancel
-        let webviewCloseButton = app.otherElements["TopBrowserBar"].buttons["Close"]
-        XCTAssertTrue(webviewCloseButton.waitForExistence(timeout: 10.0))
-        webviewCloseButton.tap()
-
-        // Tap to attempt a set up, but fail it
-        setupButton.waitForExistenceAndTap()
-        let failSetupText = app.firstDescendant(withLabel: "FAIL TEST SETUP")
-        failSetupText.waitForExistenceAndTap(timeout: 15.0)
-
-        XCTAssertTrue(app.staticTexts["The customer declined this payment."].waitForExistence(timeout: 5.0))
-
-        // Tap to attempt a set up, make it succeed
-        setupButton.waitForExistenceAndTap()
-        let approveSetupText = app.firstDescendant(withLabel: "AUTHORIZE TEST SETUP")
-        approveSetupText.waitForExistenceAndTap(timeout: 15.0)
-
-        XCTAssertTrue(app.staticTexts["Success!"].waitForExistence(timeout: 15.0))
-    }
-
-    func testCashAppPaymentMethod_setupFutureUsage() throws {
-        var settings = PaymentSheetTestPlaygroundSettings.defaultValues()
-        settings.customerMode = .new
-        settings.apmsEnabled = .on
-        settings.mode = .paymentWithSetup
-        loadPlayground(
-            app,
-            settings
-        )
-
-        app.buttons["Present PaymentSheet"].tap()
-        let payButton = app.buttons["Pay $50.99"]
-
-        // Select Cash App
-        guard let cashApp = scroll(collectionView: app.collectionViews.firstMatch, toFindCellWithId: "Cash App Pay")
-        else {
-            XCTFail()
-            return
-        }
-        cashApp.tap()
-
-        // Attempt to pay
-        payButton.tap()
-
-        // Close the webview, no need to see the successful set up
-        let webviewCloseButton = app.otherElements["TopBrowserBar"].buttons["Close"]
-        XCTAssertTrue(webviewCloseButton.waitForExistence(timeout: 10.0))
-        webviewCloseButton.tap()
-    }
-
-    func testAmazonPayPaymentMethod_setup() throws {
-        var settings = PaymentSheetTestPlaygroundSettings.defaultValues()
-        settings.customerMode = .new
-        settings.apmsEnabled = .on
-        settings.mode = .setup
-        loadPlayground(
-            app,
-            settings
-        )
-
-        app.buttons["Present PaymentSheet"].tap()
-        let setupButton = app.buttons["Set up"]
-
-        // Select Amazon Pay
-        guard let amazonPay = scroll(collectionView: app.collectionViews.firstMatch, toFindCellWithId: "Amazon Pay")
-        else {
-            XCTFail()
-            return
-        }
-        amazonPay.tap()
-
-        // Attempt set up
-        setupButton.tap()
-
-        // Close the webview, no need to see the successful set up
-        let webviewCloseButton = app.otherElements["TopBrowserBar"].buttons["Close"]
-        XCTAssertTrue(webviewCloseButton.waitForExistence(timeout: 10.0))
-        webviewCloseButton.tap()
-    }
-
-    func testAmazonPayPaymentMethod_setupFutureUsage() throws {
-        var settings = PaymentSheetTestPlaygroundSettings.defaultValues()
-        settings.customerMode = .new
-        settings.apmsEnabled = .on
-        settings.mode = .paymentWithSetup
-        loadPlayground(
-            app,
-            settings
-        )
-
-        app.buttons["Present PaymentSheet"].tap()
-        let payButton = app.buttons["Pay $50.99"]
-
-        // Select Amazon Pay
-        guard let amazonPay = scroll(collectionView: app.collectionViews.firstMatch, toFindCellWithId: "Amazon Pay")
-        else {
-            XCTFail()
-            return
-        }
-        amazonPay.tap()
-
-        // Attempt to pay
-        payButton.tap()
-
-        // Close the webview, no need to see the successful set up
-        let webviewCloseButton = app.otherElements["TopBrowserBar"].buttons["Close"]
-        XCTAssertTrue(webviewCloseButton.waitForExistence(timeout: 10.0))
-        webviewCloseButton.tap()
-    }
-
-    func testPayPalPaymentMethod_setup() throws {
-        var settings = PaymentSheetTestPlaygroundSettings.defaultValues()
-        settings.customerMode = .new
-        settings.apmsEnabled = .on
-        settings.merchantCountryCode = .FR
-        settings.currency = .eur
-        settings.mode = .setup
-        loadPlayground(
-            app,
-            settings
-        )
-
-        app.buttons["Present PaymentSheet"].tap()
-        let setupButton = app.buttons["Set up"]
-
-        // Select PayPal
-        guard let payPal = scroll(collectionView: app.collectionViews.firstMatch, toFindCellWithId: "PayPal")
-        else {
-            XCTFail()
-            return
-        }
-        payPal.tap()
-
-        XCTAssertTrue(setupButton.isEnabled)
-    }
-
-    func testPayPalPaymentMethod_setupFutureUsage() throws {
-        var settings = PaymentSheetTestPlaygroundSettings.defaultValues()
-        settings.customerMode = .new
-        settings.apmsEnabled = .on
-        settings.merchantCountryCode = .FR
-        settings.currency = .eur
-        settings.mode = .paymentWithSetup
-        loadPlayground(
-            app,
-            settings
-        )
-
-        app.buttons["Present PaymentSheet"].tap()
-        let payButton = app.buttons["Pay €50.99"]
-
-        // Select PayPal
-        guard let payPal = scroll(collectionView: app.collectionViews.firstMatch, toFindCellWithId: "PayPal")
-        else {
-            XCTFail()
-            return
-        }
-        payPal.tap()
-
-        XCTAssertTrue(payButton.isEnabled)
-    }
-
-    func testRevolutPayPaymentMethod_setup() throws {
-        var settings = PaymentSheetTestPlaygroundSettings.defaultValues()
-        settings.customerMode = .new
-        settings.apmsEnabled = .on
-        settings.merchantCountryCode = .GB
-        settings.currency = .gbp
-        settings.mode = .setup
-        loadPlayground(
-            app,
-            settings
-        )
-
-        app.buttons["Present PaymentSheet"].tap()
-        let setupButton = app.buttons["Set up"]
-
-        // Select Revolut Pay
-        guard let revolut = scroll(collectionView: app.collectionViews.firstMatch, toFindCellWithId: "Revolut Pay")
-        else {
-            XCTFail()
-            return
-        }
-        revolut.tap()
-
-        // Attempt set up
-        setupButton.tap()
-
-        // Close the webview, no need to see the successful set up
-        let webviewCloseButton = app.otherElements["TopBrowserBar"].buttons["Close"]
-        XCTAssertTrue(webviewCloseButton.waitForExistence(timeout: 10.0))
-        webviewCloseButton.tap()
-    }
-
-    func testRvolutPayPaymentMethod_setupFutureUsage() throws {
-        var settings = PaymentSheetTestPlaygroundSettings.defaultValues()
-        settings.customerMode = .new
-        settings.apmsEnabled = .on
-        settings.merchantCountryCode = .GB
-        settings.currency = .gbp
-        settings.mode = .paymentWithSetup
-        loadPlayground(
-            app,
-            settings
-        )
-
-        app.buttons["Present PaymentSheet"].tap()
-        let payButton = app.buttons["Pay £50.99"]
-
-        // Select Revolut Pay
-        guard let revolut = scroll(collectionView: app.collectionViews.firstMatch, toFindCellWithId: "Revolut Pay")
-        else {
-            XCTFail()
-            return
-        }
-        revolut.tap()
-
-        // Attempt to pay
-        payButton.tap()
-
-        // Close the webview, no need to see the successful set up
-        let webviewCloseButton = app.otherElements["TopBrowserBar"].buttons["Close"]
-        XCTAssertTrue(webviewCloseButton.waitForExistence(timeout: 10.0))
-        webviewCloseButton.tap()
-    }
-
-    func testUSBankAccountPaymentMethod() throws {
-        app.launchEnvironment = app.launchEnvironment.merging(["USE_PRODUCTION_FINANCIAL_CONNECTIONS_SDK": "false"]) { (_, new) in new }
-        var settings = PaymentSheetTestPlaygroundSettings.defaultValues()
-        settings.customerMode = .new
-        settings.apmsEnabled = .off
-        settings.allowsDelayedPMs = .on
-        loadPlayground(app, settings)
-        app.buttons["Present PaymentSheet"].tap()
-
-        // Select US Bank Account
-        guard
-            let usBankAccount = scroll(
-                collectionView: app.collectionViews.firstMatch,
-                toFindCellWithId: "US Bank Account"
-            )
-        else {
-            XCTFail()
-            return
-        }
-        usBankAccount.tap()
-
-        let continueButton = app.buttons["Continue"]
-        XCTAssertFalse(continueButton.isEnabled)
-
-        let name = app.textFields["Full name"]
-        name.tap()
-        name.typeText("John Doe")
-        name.typeText(XCUIKeyboardKey.return.rawValue)
-
-        let email = app.textFields["Email"]
-        email.tap()
-        email.typeText("test@example.com")
-        email.typeText(XCUIKeyboardKey.return.rawValue)
-
-        XCTAssertTrue(continueButton.isEnabled)
-        continueButton.tap()
-
-        let payButton = app.buttons["Pay $50.99"]
-        XCTAssertTrue(payButton.waitForExistence(timeout: 5))
-
-        let selectedMandate =
-        "By saving your bank account for Example, Inc. you agree to authorize payments pursuant to these terms."
-        let unselectedMandate = "By continuing, you agree to authorize payments pursuant to these terms."
-        XCTAssertTrue(
-            app.textViews[unselectedMandate].waitForExistence(timeout: 5)
-        )
-
-        let saveThisAccountToggle = app.switches["Save this account for future Example, Inc. payments"]
-        saveThisAccountToggle.tap()
-        XCTAssertTrue(
-            app.textViews[selectedMandate].waitForExistence(timeout: 5)
-        )
-
-        // no pay button tap because linked account is stubbed/fake in UI test
-    }
-
-    func testGrabPayPaymentMethod() throws {
-        var settings = PaymentSheetTestPlaygroundSettings.defaultValues()
-        settings.customerMode = .new // new customer
-        settings.apmsEnabled = .on
-        settings.currency = .sgd
-        settings.merchantCountryCode = .SG
-        loadPlayground(app, settings)
-
-        app.buttons["Present PaymentSheet"].tap()
-        let payButton = app.buttons["Pay SGD 50.99"]
-
-        // Select GrabPay
-        guard let grabPay = scroll(collectionView: app.collectionViews.firstMatch, toFindCellWithId: "GrabPay")
-        else {
-            XCTFail()
-            return
-        }
-        grabPay.tap()
-
-        // Attempt payment
-        payButton.tap()
-
-        // Close the webview, no need to see the successful pay
-        let webviewCloseButton = app.otherElements["TopBrowserBar"].buttons["Close"]
-        XCTAssertTrue(webviewCloseButton.waitForExistence(timeout: 10.0))
-        webviewCloseButton.tap()
-    }
-
-    func testPaymentIntent_USBankAccount() {
-        _testUSBankAccount(mode: .payment, integrationType: .normal)
-    }
-
-    func testSetupIntent_USBankAccount() {
-        _testUSBankAccount(mode: .setup, integrationType: .normal)
-    }
-
-    func testPaymentIntent_instantDebits() {
-        _testInstantDebits(mode: .payment)
-    }
-
-    func testSetupIntent_instantDebits() {
-        _testInstantDebits(mode: .setup)
-    }
-
-    func testUPIPaymentMethod() throws {
-        var settings = PaymentSheetTestPlaygroundSettings.defaultValues()
-        settings.customerMode = .new
-        settings.merchantCountryCode = .IN
-        settings.currency = .inr
-        settings.apmsEnabled = .off
-        loadPlayground(app, settings)
-
-        app.buttons["Present PaymentSheet"].waitForExistenceAndTap()
-        XCTAssertTrue(app.buttons["Pay ₹50.99"].waitForExistence(timeout: 5))
-
-        let payButton = app.buttons["Pay ₹50.99"]
-        guard let upi = scroll(collectionView: app.collectionViews.firstMatch, toFindCellWithId: "UPI") else {
-            XCTFail()
-            return
-        }
-        upi.tap()
-
-        XCTAssertFalse(payButton.isEnabled)
-        let upi_id = app.textFields["UPI ID"]
-        upi_id.tap()
-        upi_id.typeText("payment.success@stripeupi")
-        upi_id.typeText(XCUIKeyboardKey.return.rawValue)
-
-        payButton.tap()
-    }
-
-    func testUPIPaymentMethod_invalidVPA() throws {
-        var settings = PaymentSheetTestPlaygroundSettings.defaultValues()
-        settings.customerMode = .new
-        settings.merchantCountryCode = .IN
-        settings.currency = .inr
-        settings.apmsEnabled = .off
-        loadPlayground(app, settings)
-
-        app.buttons["Present PaymentSheet"].waitForExistenceAndTap()
-        XCTAssertTrue(app.buttons["Pay ₹50.99"].waitForExistence(timeout: 5))
-
-        let payButton = app.buttons["Pay ₹50.99"]
-        guard let upi = scroll(collectionView: app.collectionViews.firstMatch, toFindCellWithId: "UPI") else {
-            XCTFail()
-            return
-        }
-        upi.tap()
-
-        XCTAssertFalse(payButton.isEnabled)
-        let upi_id = app.textFields["UPI ID"]
-        upi_id.tap()
-        upi_id.typeText("payment.success")
-        upi_id.typeText(XCUIKeyboardKey.return.rawValue)
-
-        XCTAssertFalse(payButton.isEnabled)
-    }
-
-    // MARK: Card brand choice
-    func testCardBrandChoice() throws {
-        // Currently only our French merchant is eligible for card brand choice
-        var settings = PaymentSheetTestPlaygroundSettings.defaultValues()
-        settings.customerMode = .new
-        settings.merchantCountryCode = .FR
-        settings.currency = .eur
-        settings.preferredNetworksEnabled = .off
-        loadPlayground(app, settings)
-
-        _testCardBrandChoice(settings: settings)
-    }
-
-    func testCardBrandChoice_setup() throws {
-        // Currently only our French merchant is eligible for card brand choice
-        var settings = PaymentSheetTestPlaygroundSettings.defaultValues()
-        settings.mode = .setup
-        settings.customerMode = .new
-        settings.merchantCountryCode = .FR
-        settings.currency = .eur
-        settings.preferredNetworksEnabled = .off
-        loadPlayground(app, settings)
-
-        _testCardBrandChoice(isSetup: true, settings: settings)
-    }
-
-    func testCardBrandChoice_deferred() throws {
-        // Currently only our French merchant is eligible for card brand choice
-        var settings = PaymentSheetTestPlaygroundSettings.defaultValues()
-        settings.customerMode = .new
-        settings.merchantCountryCode = .FR
-        settings.currency = .eur
-        settings.preferredNetworksEnabled = .off
-        settings.integrationType = .deferred_csc
-        loadPlayground(app, settings)
-
-        _testCardBrandChoice(settings: settings)
-    }
-
-    func testCardBrandChoiceWithPreferredNetworks() throws {
-        // Currently only our French merchant is eligible for card brand choice
-        var settings = PaymentSheetTestPlaygroundSettings.defaultValues()
-        settings.customerMode = .new
-        settings.merchantCountryCode = .FR
-        settings.currency = .eur
-        settings.preferredNetworksEnabled = .on
-        loadPlayground(app, settings)
-
-        app.buttons["Present PaymentSheet"].tap()
-
-        // We should have selected Visa due to preferreedNetworks configuration API
-        let cardBrandTextField = app.textFields["Visa"]
-        let cardBrandChoiceDropdown = app.pickerWheels.firstMatch
-        // Card brand choice textfield/dropdown should not be visible
-        XCTAssertFalse(cardBrandTextField.waitForExistence(timeout: 2))
-
-        let numberField = app.textFields["Card number"]
-        numberField.tap()
-        // Enter 8 digits to start fetching card brand
-        numberField.typeText("49730197")
-
-        // Card brand choice drop down should be enabled
-        cardBrandTextField.tap()
-        XCTAssertTrue(cardBrandChoiceDropdown.waitForExistence(timeout: 5))
-        cardBrandChoiceDropdown.swipeDown()
-        app.toolbars.buttons["Cancel"].tap()
-
-        // We should have selected Visa due to preferreedNetworks configuration API
-        XCTAssertTrue(app.textFields["Visa"].waitForExistence(timeout: 2))
-
-        // Clear card text field, should reset selected card brand
-        numberField.tap()
-        numberField.clearText()
-
-        // We should reset to showing unknown in the textfield for card brand
-        XCTAssertFalse(app.textFields["Select card brand (optional)"].waitForExistence(timeout: 2))
-
-        // Type full card number to start fetching card brands again
-        numberField.forceTapWhenHittableInTestCase(self)
-        app.typeText("4000002500001001")
-        app.textFields["expiration date"].waitForExistenceAndTap(timeout: 5.0)
-        app.typeText("1228") // Expiry
-        app.typeText("123") // CVC
-        app.toolbars.buttons["Done"].tap() // Country picker toolbar's "Done" button
-        app.typeText("12345") // Postal
-
-        // Card brand choice drop down should be enabled and we should auto select Visa
-        XCTAssertTrue(app.textFields["Visa"].waitForExistence(timeout: 5))
-
-        // Finish checkout
-        app.buttons["Pay €50.99"].tap()
-        let successText = app.staticTexts["Success!"]
-        XCTAssertTrue(successText.waitForExistence(timeout: 10.0))
-    }
-
-    func testCardBrandChoiceSavedCard() {
-        // Currently only our French merchant is eligible for card brand choice
-        var settings = PaymentSheetTestPlaygroundSettings.defaultValues()
-        settings.customerMode = .new
-        settings.merchantCountryCode = .FR
-        settings.currency = .eur
-        loadPlayground(app, settings)
-
-        app.buttons["Present PaymentSheet"].waitForExistenceAndTap(timeout: 5)
-        let numberField = app.textFields["Card number"]
-        let cardBrandChoiceDropdown = app.pickerWheels.firstMatch
-
-        // Type full card number to start fetching card brands again
-        numberField.forceTapWhenHittableInTestCase(self)
-        app.typeText("4000002500001001")
-        app.textFields["expiration date"].waitForExistenceAndTap(timeout: 5.0)
-        app.typeText("1228") // Expiry
-        app.typeText("123") // CVC
-        app.toolbars.buttons["Done"].tap() // Country picker toolbar's "Done" button
-        app.typeText("12345") // Postal
-
-        // Card brand choice drop down should be enabled
-        XCTAssertTrue(app.textFields["Select card brand (optional)"].waitForExistenceAndTap(timeout: 5))
-        XCTAssertTrue(cardBrandChoiceDropdown.waitForExistence(timeout: 5))
-        cardBrandChoiceDropdown.selectNextOption()
-        app.toolbars.buttons["Done"].tap()
-
-        // We should have selected cartes bancaires
-        XCTAssertTrue(app.textFields["Cartes Bancaires"].waitForExistence(timeout: 5))
-
-        // toggle save this card on
-        let saveThisCardToggle = app.switches["Save this card for future Example, Inc. payments"]
-        saveThisCardToggle.tap()
-        XCTAssertTrue(saveThisCardToggle.isSelected)
-
-        // Finish checkout
-        app.buttons["Pay €50.99"].tap()
-        let successText = app.staticTexts["Success!"]
-        XCTAssertTrue(successText.waitForExistence(timeout: 10.0))
-
-        // Reload w/ same customer
-        reload(app, settings: settings)
-        app.buttons["Present PaymentSheet"].waitForExistenceAndTap(timeout: 5)
-        // Saved card should show the cartes bancaires logo
-        XCTAssertTrue(app.staticTexts["••••1001"].waitForExistence(timeout: 5.0))
-        XCTAssertTrue(app.images["carousel_card_cartes_bancaires"].waitForExistence(timeout: 5))
-
-        let editButton = app.staticTexts["Edit"]
-        XCTAssertTrue(editButton.waitForExistence(timeout: 60.0))
-        editButton.tap()
-
-        // Saved card should show the edit icon since it is co-branded
-        XCTAssertTrue(app.buttons["CircularButton.Edit"].waitForExistenceAndTap(timeout: 5))
-
-        // Update this card
-        XCTAssertTrue(app.textFields["Cartes Bancaires"].waitForExistenceAndTap(timeout: 5))
-        XCTAssertTrue(app.pickerWheels.firstMatch.waitForExistence(timeout: 5))
-        app.pickerWheels.firstMatch.swipeUp()
-        app.toolbars.buttons["Done"].tap()
-        app.buttons["Update"].waitForExistenceAndTap(timeout: 5)
-
-        // We should have updated to Visa
-        XCTAssertTrue(app.images["carousel_card_visa"].waitForExistence(timeout: 5))
-
-        // Update this card again
-        XCTAssertTrue(app.buttons["CircularButton.Edit"].waitForExistenceAndTap(timeout: 5))
-        XCTAssertTrue(app.textFields["Visa"].waitForExistenceAndTap(timeout: 5))
-        XCTAssertTrue(app.pickerWheels.firstMatch.waitForExistence(timeout: 5))
-        app.pickerWheels.firstMatch.swipeDown()
-        app.toolbars.buttons["Done"].tap()
-        app.buttons["Update"].waitForExistenceAndTap(timeout: 5)
-
-        // We should have updated to Cartes Bancaires
-        XCTAssertTrue(app.images["carousel_card_cartes_bancaires"].waitForExistence(timeout: 5))
-        app.buttons["Done"].tap()
-
-        // Pay with this card
-        app.buttons["Pay €50.99"].tap()
-        XCTAssertTrue(app.staticTexts["Success!"].waitForExistence(timeout: 10.0))
-
-        // Reload w/ same customer
-        reload(app, settings: settings)
-        app.buttons["Present PaymentSheet"].waitForExistenceAndTap(timeout: 5)
-        // Saved card should show the cartes bancaires logo
-        XCTAssertTrue(app.staticTexts["••••1001"].waitForExistence(timeout: 5.0))
-        XCTAssertTrue(app.images["carousel_card_cartes_bancaires"].waitForExistence(timeout: 5))
-
-        // Remove this card
-        XCTAssertTrue(app.staticTexts["Edit"].waitForExistenceAndTap(timeout: 60.0))
-        XCTAssertTrue(app.buttons["CircularButton.Edit"].waitForExistenceAndTap(timeout: 5))
-        XCTAssertTrue(app.buttons["Remove card"].waitForExistenceAndTap(timeout: 5))
-        let confirmRemoval = app.alerts.buttons["Remove"]
-        XCTAssertTrue(confirmRemoval.waitForExistence(timeout: 5))
-        confirmRemoval.tap()
-
-        // Card should be removed
-        XCTAssertFalse(app.staticTexts["••••1001"].waitForExistence(timeout: 5.0))
-    }
-
-    // This only tests the PaymentSheet + PaymentIntent flow.
-    // Other confirmation flows are tested in PaymentSheet+LPMTests.swift
-    func testSEPADebitPaymentMethod_PaymentSheet() {
-        var settings = PaymentSheetTestPlaygroundSettings.defaultValues()
-        settings.currency = .eur
-        settings.allowsDelayedPMs = .on
-        loadPlayground(app, settings)
-        app.buttons["Present PaymentSheet"].tap()
-
-        guard let sepa = scroll(collectionView: app.collectionViews.firstMatch, toFindCellWithId: "SEPA Debit") else { XCTFail("Couldn't find SEPA"); return; }
-        sepa.tap()
-
-        app.textFields["Full name"].tap()
-        app.typeText("John Doe" + XCUIKeyboardKey.return.rawValue)
-        app.typeText("test@example.com" + XCUIKeyboardKey.return.rawValue)
-        app.typeText("AT611904300234573201" + XCUIKeyboardKey.return.rawValue)
-        app.textFields["Address line 1"].tap()
-        app.typeText("510 Townsend St" + XCUIKeyboardKey.return.rawValue)
-        app.typeText("Floor 3" + XCUIKeyboardKey.return.rawValue)
-        app.typeText("San Francisco" + XCUIKeyboardKey.return.rawValue)
-        app.textFields["ZIP"].tap()
-        app.typeText("94102" + XCUIKeyboardKey.return.rawValue)
-        app.buttons["Pay €50.99"].tap()
-        let successText = app.staticTexts["Success!"]
-        XCTAssertTrue(successText.waitForExistence(timeout: 10.0))
-    }
-
-    func testAlipayPaymentMethod() throws {
-        var settings = PaymentSheetTestPlaygroundSettings.defaultValues()
-        settings.customerMode = .new
-        settings.merchantCountryCode = .US
-        settings.currency = .usd
-        settings.apmsEnabled = .on
-        loadPlayground(app, settings)
-
-        app.buttons["Present PaymentSheet"].tap()
-
-        let payButton = app.buttons["Pay $50.99"]
-        guard let alipay = scroll(collectionView: app.collectionViews.firstMatch, toFindCellWithId: "Alipay") else {
-            XCTFail()
-            return
-        }
-        alipay.tap()
-
-        XCTAssertTrue(payButton.isEnabled)
-        payButton.tap()
-
-        let approvePaymentText = app.firstDescendant(withLabel: "AUTHORIZE TEST PAYMENT")
-        approvePaymentText.waitForExistenceAndTap(timeout: 15.0)
-
-        XCTAssertTrue(app.staticTexts["Success!"].waitForExistence(timeout: 15.0))
-    }
-
-    func testOXXOPaymentMethod() throws {
-        var settings = PaymentSheetTestPlaygroundSettings.defaultValues()
-        settings.customerMode = .new
-        settings.merchantCountryCode = .MX
-        settings.currency = .mxn
-        settings.apmsEnabled = .off
-        settings.allowsDelayedPMs = .on
-        loadPlayground(app, settings)
-
-        app.buttons["Present PaymentSheet"].tap()
-
-        guard let oxxo = scroll(collectionView: app.collectionViews.firstMatch, toFindCellWithId: "OXXO") else {
-            XCTFail()
-            return
-        }
-        oxxo.tap()
-
-        let name = app.textFields["Full name"]
-        name.tap()
-        name.typeText("Jane Doe")
-        name.typeText(XCUIKeyboardKey.return.rawValue)
-
-        let email = app.textFields["Email"]
-        email.tap()
-        email.typeText("foo@bar.com")
-        email.typeText(XCUIKeyboardKey.return.rawValue)
-
-        let payButton = app.buttons["Pay MX$50.99"]
-        XCTAssertTrue(payButton.isEnabled)
-        payButton.tap()
-
-        // Just check that a web view exists after tapping buy.
-        let webviewCloseButton = app.otherElements["TopBrowserBar"].buttons["Close"]
-        XCTAssertTrue(webviewCloseButton.waitForExistence(timeout: 10.0))
-        webviewCloseButton.tap()
-
-        XCTAssertTrue(app.staticTexts["Success!"].waitForExistence(timeout: 15.0))
-    }
-
-    func testBoletoPaymentMethod() throws {
-        var settings = PaymentSheetTestPlaygroundSettings.defaultValues()
-        settings.customerMode = .new
-        settings.merchantCountryCode = .BR
-        settings.currency = .brl
-        settings.apmsEnabled = .off
-        settings.allowsDelayedPMs = .on
-        loadPlayground(app, settings)
-
-        app.buttons["Present PaymentSheet"].tap()
-
-        guard let boleto = scroll(collectionView: app.collectionViews.firstMatch, toFindCellWithId: "Boleto") else {
-            XCTFail()
-            return
-        }
-        boleto.tap()
-
-        let name = app.textFields["Full name"]
-        name.tap()
-        app.typeText("Jane Doe")
-        app.typeText(XCUIKeyboardKey.return.rawValue)
-        app.typeText("foo@bar.com")
-        app.typeText(XCUIKeyboardKey.return.rawValue)
-        app.typeText("00000000000")
-        app.typeText(XCUIKeyboardKey.return.rawValue)
-        app.typeText("123 fake st")
-        app.typeText(XCUIKeyboardKey.return.rawValue)
-        app.typeText(XCUIKeyboardKey.return.rawValue)
-        app.typeText("City")
-        app.typeText(XCUIKeyboardKey.return.rawValue)
-        app.typeText("AC")  // Valid brazilian state code.
-        app.typeText(XCUIKeyboardKey.return.rawValue)
-        app.typeText("11111111")
-        app.typeText(XCUIKeyboardKey.return.rawValue)
-
-        let payButton = app.buttons["Pay R$50.99"]
-        XCTAssertTrue(payButton.isEnabled)
-        payButton.tap()
-
-        // Just check that a web view exists after tapping buy.
-        let webviewCloseButton = app.otherElements["TopBrowserBar"].buttons["Close"]
-        XCTAssertTrue(webviewCloseButton.waitForExistence(timeout: 10.0))
-        webviewCloseButton.tap()
-
-        XCTAssertTrue(app.staticTexts["Success!"].waitForExistence(timeout: 15.0))
-    }
-
-    func testPayNowPaymentMethod() throws {
-        var settings = PaymentSheetTestPlaygroundSettings.defaultValues()
-        settings.customerMode = .new // new customer
-        settings.apmsEnabled = .on
-        settings.currency = .sgd
-        settings.merchantCountryCode = .SG
-        loadPlayground(app, settings)
-
-        app.buttons["Present PaymentSheet"].tap()
-        let payButton = app.buttons["Pay SGD 50.99"]
-
-        // Select PayNow
-        guard let payNow = scroll(collectionView: app.collectionViews.firstMatch, toFindCellWithId: "PayNow")
-        else {
-            XCTFail()
-            return
-        }
-        payNow.tap()
-
-        // Attempt payment
-        payButton.tap()
-
-        // Close the webview, no need to see the successful pay
-        let webviewCloseButton = app.otherElements["TopBrowserBar"].buttons["Close"]
-        XCTAssertTrue(webviewCloseButton.waitForExistence(timeout: 10.0))
-        webviewCloseButton.tap()
-    }
-
-    func testPromptPayPaymentMethod() throws {
-        var settings = PaymentSheetTestPlaygroundSettings.defaultValues()
-        settings.customerMode = .new // new customer
-        settings.apmsEnabled = .on
-        settings.currency = .thb
-        settings.merchantCountryCode = .TH
-        loadPlayground(app, settings)
-
-        app.buttons["Present PaymentSheet"].tap()
-
-        // Select PromptPay
-        guard let promptPay = scroll(collectionView: app.collectionViews.firstMatch, toFindCellWithId: "PromptPay")
-        else {
-            XCTFail()
-            return
-        }
-        promptPay.tap()
-
-        // Fill in email
-        let email = app.textFields["Email"]
-        email.tap()
-        email.typeText("foo@bar.com")
-        email.typeText(XCUIKeyboardKey.return.rawValue)
-
-        // Attempt payment
-        XCTAssertTrue(app.buttons["Pay THB 50.99"].waitForExistenceAndTap(timeout: 5.0))
-
-        // Close the webview, no need to see the successful pay
-        let webviewCloseButton = app.otherElements["TopBrowserBar"].buttons["Close"]
-        XCTAssertTrue(webviewCloseButton.waitForExistence(timeout: 10.0))
-        webviewCloseButton.tap()
-    }
-
-    func testSwishPaymentMethod() throws {
-        var settings = PaymentSheetTestPlaygroundSettings.defaultValues()
-        settings.customerMode = .new // new customer
-        settings.apmsEnabled = .off
-        settings.currency = .sek
-        settings.merchantCountryCode = .FR
-        loadPlayground(app, settings)
-
-        app.buttons["Present PaymentSheet"].tap()
-
-        // Select Swish
-        guard let swish = scroll(collectionView: app.collectionViews.firstMatch, toFindCellWithId: "Swish")
-        else {
-            XCTFail()
-            return
-        }
-        swish.tap()
-
-        // Attempt payment
-        XCTAssertTrue(app.buttons["Pay SEK 50.99"].waitForExistenceAndTap(timeout: 5.0))
-
-        let approvePaymentText = app.firstDescendant(withLabel: "AUTHORIZE TEST PAYMENT")
-        approvePaymentText.waitForExistenceAndTap(timeout: 15.0)
-
-        XCTAssertTrue(app.staticTexts["Success!"].waitForExistence(timeout: 15.0))
-    }
-
-    func testSavedSEPADebitPaymentMethod_FlowController_ShowsMandate() {
-        var settings = PaymentSheetTestPlaygroundSettings.defaultValues()
-        settings.uiStyle = .flowController
-        settings.customerMode = .new
-        settings.applePayEnabled = .off // disable Apple Pay
-        settings.mode = .setup
-        settings.allowsDelayedPMs = .on
-        loadPlayground(app, settings)
-
-        let paymentMethodButton = app.buttons["Payment method"]
-        XCTAssertTrue(paymentMethodButton.waitForExistence(timeout: 60.0))
-        paymentMethodButton.tap()
-
-        // Save SEPA
-        app.buttons["+ Add"].waitForExistenceAndTap()
-        guard let sepa = scroll(collectionView: app.collectionViews.firstMatch, toFindCellWithId: "SEPA Debit") else {
-            XCTFail("Couldn't find SEPA")
-            return
-        }
-        sepa.tap()
-
-        app.textFields["Full name"].tap()
-        app.typeText("John Doe" + XCUIKeyboardKey.return.rawValue)
-        app.typeText("test@example.com" + XCUIKeyboardKey.return.rawValue)
-        app.typeText("AT611904300234573201" + XCUIKeyboardKey.return.rawValue)
-        app.textFields["Address line 1"].tap()
-        app.typeText("510 Townsend St" + XCUIKeyboardKey.return.rawValue)
-        app.typeText("Floor 3" + XCUIKeyboardKey.return.rawValue)
-        app.typeText("San Francisco" + XCUIKeyboardKey.return.rawValue)
-        app.textFields["ZIP"].tap()
-        app.typeText("94102" + XCUIKeyboardKey.return.rawValue)
-        app.buttons["Continue"].tap()
-        app.buttons["Confirm"].tap()
-        XCTAssertTrue(app.staticTexts["Success!"].waitForExistence(timeout: 10.0))
-
-        // Reload w/ same customer
-        reload(app, settings: settings)
-        // This time, expect SEPA to be pre-selected as the default
-        XCTAssert(paymentMethodButton.label.hasPrefix("••••3201, sepa_debit"))
-
-        // Tapping confirm without presenting flowcontroller should show the mandate
-        app.buttons["Confirm"].tap()
-        XCTAssertTrue(app.otherElements.matching(identifier: "mandatetextview").element.waitForExistence(timeout: 1))
-        // Tapping out should cancel the payment
-        app.tap()
-        XCTAssertTrue(app.staticTexts["Payment canceled."].waitForExistence(timeout: 10.0))
-        // Tapping confirm again and hitting continue should confirm the payment
-        app.buttons["Confirm"].tap()
-        app.buttons["Continue"].tap()
-        XCTAssertTrue(app.staticTexts["Success!"].waitForExistence(timeout: 10.0))
-
-        // Reload w/ same customer
-        reload(app, settings: settings)
-        // If you present the flowcontroller and see the mandate...
-        XCTAssert(paymentMethodButton.label.hasPrefix("••••3201, sepa_debit"))
-        paymentMethodButton.waitForExistenceAndTap()
-
-        XCTAssertTrue(app.otherElements.matching(identifier: "mandatetextview").element.exists)
-        // ...you shouldn't see the mandate again when you confirm
-        app.buttons["Continue"].tap()
-        app.buttons["Confirm"].tap()
-        XCTAssertTrue(app.staticTexts["Success!"].waitForExistence(timeout: 10.0))
-    }
-
-    func testMultibancoPaymentMethod() throws {
-        var settings = PaymentSheetTestPlaygroundSettings.defaultValues()
-        settings.customerMode = .new
-        settings.currency = .eur
-        settings.apmsEnabled = .off
-        settings.allowsDelayedPMs = .on
-        loadPlayground(app, settings)
-
-        app.buttons["Present PaymentSheet"].tap()
-
-        guard let multibanco = scroll(collectionView: app.collectionViews.firstMatch, toFindCellWithId: "Multibanco") else {
-            XCTFail()
-            return
-        }
-        multibanco.tap()
-
-        let email = app.textFields["Email"]
-        email.tap()
-        app.typeText("foo@bar.com")
-        app.typeText(XCUIKeyboardKey.return.rawValue)
-
-        let payButton = app.buttons["Pay €50.99"]
-        XCTAssertTrue(payButton.isEnabled)
-        payButton.tap()
-
-        // Just check that a web view exists after tapping buy.
-        let webviewCloseButton = app.otherElements["TopBrowserBar"].buttons["Close"]
-        XCTAssertTrue(webviewCloseButton.waitForExistence(timeout: 10.0))
-        webviewCloseButton.tap()
-
-        XCTAssertTrue(app.staticTexts["Success!"].waitForExistence(timeout: 15.0))
-    }
-}
-
 class PaymentSheetDeferredUITests: PaymentSheetUITestCase {
 
     // MARK: Deferred tests (client-side)
